refactor: null 타입 체크
