+++ frontend/src/hooks/useAnimateClickedPin.ts
@@ -1,13 +1,13 @@
 import { useEffect } from 'react';
 
-const useAnimateClickedPin = (map: TMap, markers: Marker[]) => {
+const useAnimateClickedPin = (map: TMap | null, markers: Marker[]) => {
   useEffect(() => {
     const queryParams = new URLSearchParams(location.search);
     if (queryParams.has('pinDetail')) {
       const pinId = queryParams.get('pinDetail');
 
       const marker = markers.find((marker: Marker) => marker.id === pinId);
-      if (marker) {
+      if (marker && map) {
         map.setCenter(marker.getPosition());
         map.setZoom(17);
       }
