+++ frontend/src/pages/NewTopic.tsx
@@ -5,7 +5,7 @@ import Space from '../components/common/Space';
 import Button from '../components/common/Button';
 import { postApi } from '../apis/postApi';
 import useNavigator from '../hooks/useNavigator';
-import { NewTopicFormProps } from '../types/FormValues';
+import { NewTopicFormProps } from '../types/tmap';
 import useFormValues from '../hooks/useFormValues';
 import { useLocation } from 'react-router-dom';
 import useToast from '../hooks/useToast';
