+++ frontend/src/pages/UpdatedPinDetail.tsx
@@ -4,7 +4,7 @@ import Text from '../components/common/Text';
 import Box from '../components/common/Box';
 import { putApi } from '../apis/putApi';
 import { SetURLSearchParams } from 'react-router-dom';
-import { ModifyPinFormProps } from '../types/FormValues';
+import { ModifyPinFormProps } from '../types/tmap';
 import InputContainer from '../components/InputContainer';
 import { hasErrorMessage, hasNullValue } from '../validations';
 import useToast from '../hooks/useToast';
