+++ frontend/src/types/tmap.d.ts
