--- frontend/src/utils/narrowNotNull.ts
@@ -1,5 +1,0 @@
-type NotNull<T> = T extends null | undefined ? never : T;
-
-function narrowNotNull<T>(value: T): NotNull<T> | null {
-  return value !== null && value !== undefined ? (value as NotNull<T>) : null;
-}
